Add parsers w/tests
+ Add parser for rebar
+ Add parser for OTP .app files
+ Add parser for Edam .profile files
+ Resolve packages dynamicly instead of pre-generating dependency tree
 + Add constraints to resolver
+ Add in-memory catalog
+ Add local catalog
+ Add more versatile parser logic
 + Support for applying options pr. catalog
 + Allow matching against only a subset of the catalog modules
+ Port existing eunit tests in edm_cfg_server to common tests

Resolving packages:
Dependencies are now defined as a triplet (Name, Constraints, Opts),
when the parser returns, the runtime tries to resolve all these triplets
to actual #pkg{} record by matching Name and Constraints with the
package templates found in the indiviudual catalogs; on match Opts
are applied to the package.

Each command module must handle unmet dependencies in a way the developer
seems fit.